Extract tag components
add: create TagButton component for consistent tag rendering
refactor: move GetContrastingColourForHSL to globals module
refactor: extract tag rendering logic into TagButton component
