+++ app/renderer/src/components/AppMenuModal.tsx
@@ -14,6 +14,7 @@ import {
     setDisplayedModal,
     appSettings,
     systemFonts,
+    GetContrastingColourForHSL,
 } from '../modules/globals'
 import {
     ClearWebsiteCache,
@@ -353,27 +354,6 @@ const AppearanceSettingsView: Component = () => {
         cssString = !valuesOnly ? 'hsl(' + cssString + ')' : cssString
 
         return cssString
-    }
-
-    const GetContrastingColourForHSL = (hslColour: string) => {
-        const match = hslColour.match(/\d+/g)
-
-        if (!match || match.length < 3) {
-            console.log('Invalid HSL:', hslColour)
-            return hslColour
-        }
-
-        let [h, s, l] = match.map(Number)
-
-        const contrastingHue = (h + 180) % 360
-
-        if (l <= 50) {
-            l = 90
-        } else {
-            l = 10
-        }
-
-        return `hsl(${contrastingHue}, ${s}%, ${l}%)`
     }
 
     return (
